Modified scripts to run on Ingerid-PC.
The postProcess_MM.py seems to not have been run on Ingerid-PC before, and some minor changes were required to achieve this.
Most notably, I was not able to build Slicer in order to take advantage of the BRAINSTools library, which is included there, so I resorted to building the BRAINSTools library separatly.