Modified the script to work on Ingerid-PC and addad function to update label ima$
The main problem when moving to Ingerid-PC was to convert images from
nrrd to NIFTI. I could not get the CLI modules in Slicer to build
properly, so I built the BRAINSTools separately. However, the DWIConvert
that had been used for converting previously did still not work, so I
ended up u$

I also added a function to replace the existing label images and apply
the existing transformation to it.